+++ taichi/program/program.h
@@ -24,9 +24,7 @@
 #include "taichi/program/snode_rw_accessors_bank.h"
 #include "taichi/program/context.h"
 #include "taichi/struct/snode_tree.h"
-#include "taichi/system/memory_pool.h"
 #include "taichi/system/threading.h"
-#include "taichi/system/unified_allocator.h"
 #include "taichi/program/sparse_matrix.h"
 #include "taichi/ir/mesh.h"
