+++ taichi/rhi/cuda/cuda_device.cpp
@@ -1,5 +1,5 @@
 #include "taichi/rhi/cuda/cuda_device.h"
-#include "taichi/system/memory_pool.h"
+#include "taichi/rhi/llvm/memory_pool.h"
 
 namespace taichi::lang {
