+++ taichi/runtime/llvm/runtime_module/runtime.cpp
@@ -284,7 +284,6 @@ struct PhysicalCoordinates {
 STRUCT_FIELD_ARRAY(PhysicalCoordinates, val);
 
 #include "taichi/program/context.h"
-#include "taichi/runtime/llvm/runtime_module/mem_request.h"
 
 STRUCT_FIELD_ARRAY(RuntimeContext, grad_args);
 STRUCT_FIELD(RuntimeContext, runtime);
