+++ taichi/runtime/program_impls/llvm/llvm_program.h
@@ -6,7 +6,6 @@
 #include "taichi/runtime/llvm/llvm_offline_cache.h"
 #include "taichi/program/compile_config.h"
 #include "taichi/runtime/llvm/llvm_runtime_executor.h"
-#include "taichi/system/memory_pool.h"
 #include "taichi/program/program_impl.h"
 #include "taichi/program/parallel_executor.h"
 #include "taichi/util/bit.h"
