+++ taichi/runtime/program_impls/vulkan/vulkan_program.h
@@ -11,7 +11,6 @@
 #include "taichi/runtime/gfx/snode_tree_manager.h"
 #include "taichi/rhi/vulkan/vulkan_device.h"
 
-#include "taichi/system/memory_pool.h"
 #include "taichi/common/logging.h"
 #include "taichi/struct/snode_tree.h"
 #include "taichi/program/snode_expr_utils.h"
