+++ tests/cpp/aot/gfx_utils.cpp
@@ -1,6 +1,7 @@
 #include "tests/cpp/aot/gfx_utils.h"
 
 #include "taichi/runtime/gfx/aot_module_loader_impl.h"
+#include "taichi/rhi/llvm/memory_pool.h"
 
 namespace taichi::lang {
 namespace aot_test_utils {
