+++ tests/cpp/backends/dx11_device_test.cpp
@@ -6,7 +6,6 @@
 #include "taichi/rhi/dx/dx_device.h"
 #include "taichi/rhi/dx/dx_info_queue.h"
 #include "taichi/runtime/program_impls/dx/dx_program.h"
-#include "taichi/system/memory_pool.h"
 #include "tests/cpp/program/test_program.h"
 
 namespace taichi::lang {
